remove the gamma_law EOS and rename gamma_law_general to gamma_law (#653)
we no longer need the simpler version of this EOS since we have templated EOSes
and can define types that carry a reduced set of information.
