--- EOS/gamma_law_general/Make.package
@@ -1,6 +1,0 @@
-ifeq ($(USE_FORT_EOS),TRUE)
-  F90EXE_sources += actual_eos.F90
-endif
-ifeq ($(USE_CXX_EOS),TRUE)
-  CEXE_headers += actual_eos.H
-endif
