--- EOS/gamma_law_general/_parameters
@@ -1,4 +1,0 @@
-@namespace: eos
-
-eos_gamma            real     5.d0/3.d0
-eos_assume_neutral   logical  .true.
