+++ networks/Make.package
@@ -14,7 +14,7 @@ ifeq ($(USE_REACT), TRUE)
 
   ifeq ($(INTEGRATOR_DIR), VODE90)
     override INTEGRATOR_DIR = VODE
-    $(warn VODE90 has been renamed VODE)
+    $(warning VODE90 has been renamed VODE)
   endif
 
   ifeq ($(INTEGRATOR_DIR), VODE)
